Use singleReference to keep wrapped handler reference. Create new event with correct type to pass to handler. Renamed file. Fixed whitespace.
